Refactor: Fixed mix of TS and JS on Catalog.js
The project wasn't able to follow it's path due to a mix of code in the proyect, back-end was fully written in JS while front-end had TS syntax. The complete project has to be JS only so it's important that the whole project moves from the TS syntax to a JS syntax. 
See: https://github.com/ichirodev/travelia/issues/1#issue-925705822

travelia-app/src/app/Catalog.js does not have this problem anymore.

this commit also change the model for places adding 'cost' and 'image' to the object.